# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVLT Order Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface FormItem {
  id: string;
  name: string;
  price: number;
  category?: string | null;
  maxQuantity?: number;
}

export interface Form {
  id: string;
  title: string;
  description?: string | null;
  slug?: string | null;
  isPublished: boolean;
  items: FormItem[];
  deliveryMode: string;
  paymentMethod: string;
  orderDeadline?: string | null;
  pickupTime?: string | null;
  pickupLocation?: string | null;
  googleSheetId?: string | null;
  googleSheetEmail?: string | null;
  googleSheetUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFormRequest {
  title: string;
  description?: string | null;
  items?: FormItem[];
  deliveryMode?: string;
  paymentMethod?: string;
  orderDeadline?: string | null;
  pickupTime?: string | null;
  pickupLocation?: string | null;
  googleSheetEmail?: string | null;
}

export interface UpdateFormRequest {
  title?: string;
  description?: string | null;
  items?: FormItem[];
  deliveryMode?: string | null;
  paymentMethod?: string | null;
  orderDeadline?: string | null;
  pickupTime?: string | null;
  pickupLocation?: string | null;
  googleSheetEmail?: string | null;
}

export interface OrderItem {
  itemId: string;
  itemName: string;
  price: number;
  quantity: number;
  total: number;
}

export interface SubmitOrderRequest {
  customerName: string;
  phone: string;
  address: string;
  items: OrderItem[];
  totalItems: number;
  totalAmount: number;
}

export interface Submission {
  id: string;
  formId: string;
  customerName: string;
  phone: string;
  address: string;
  items: OrderItem[];
  totalItems: number;
  totalAmount: number;
  paymentMethod: string;
  deliveryMode: string;
  createdAt: string;
}