# lib/api-zod/src/generated/types/createFormRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVLT Order Management API
 * OpenAPI spec version: 0.1.0
 */
import type { FormItem } from "./formItem";

export interface CreateFormRequest {
  title: string;
  description?: string | null;
  items?: FormItem[];
  deliveryMode?: string;
  paymentMethod?: string;
  orderDeadline?: string | null;
  pickupTime?: string | null;
  pickupLocation?: string | null;
  googleSheetEmail?: string | null;
}