# lib/api-zod/src/generated/types/form.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVLT Order Management API
 * OpenAPI spec version: 0.1.0
 */
import type { FormItem } from "./formItem";

export interface Form {
  id: string;
  title: string;
  description?: string | null;
  slug?: string | null;
  isPublished: boolean;
  items: FormItem[];
  deliveryMode: string;
  paymentMethod: string;
  orderDeadline?: string | null;
  pickupTime?: string | null;
  pickupLocation?: string | null;
  googleSheetId?: string | null;
  googleSheetEmail?: string | null;
  googleSheetUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}